refactor: 각 탭별 컴포넌트 구조 분리 및 폴더 구조 리팩토링 #97
